netlify/functions/api.py
# Flask API converted to Netlify Functions
import openai
import os
from flask import Flask, jsonify, request
from flask_cors import CORS
import json
import PyPDF2
import io
import numpy as np
import pandas as pd
import uuid
from datetime import datetime
from sklearn.ensemble import RandomForestClassifier
from dotenv import load_dotenv

netlify_handler = None

app = Flask(__name__)
load_dotenv()

@app.route("/.netlify/functions/api/applications", methods=["GET"])
def get_applications():
    sample_applications = get_sample_data()
    return jsonify(sample_applications)


@app.route("/.netlify/functions/api/applications", methods=["POST"])
def upload_application():
    try:
        application_text = ""
        if "file" in request.files:
            file = request.files["file"]
            application_text = extract_text_from_pdf(file)
        else:
            data = request.get_json()
            application_text = data.get("text", "")

        if not application_text.strip():
            return jsonify({"error": "No application text provided"}), 400

        ai_analysis = analyze_application_with_ai(application_text)
        risk_score = calculate_risk_score(ai_analysis)
        eligibility_score = calculate_eligibility_score(ai_analysis)
        completeness_score = calculate_completeness_score(ai_analysis)

        new_application = {
            "id": str(uuid.uuid4()),
            "organization": "Uploaded Organization",
            "project_title": "Uploaded Project",
            "requested_amount": 0,
            "category": "General",
            "neighborhood": "Toronto",
            "risk_score": risk_score,
            "eligibility_score": eligibility_score,
            "completeness_score": completeness_score,
            "status": "Under Review",
            "submitted_date": datetime.now().strftime("%Y-%m-%d"),
            "ai_analysis": ai_analysis,
        }

        sample_applications = get_sample_data()
        sample_applications.append(new_application)

        return jsonify(new_application), 201
    except Exception as e:
        return jsonify({"error": str(e)}), 500


@app.route("/.netlify/functions/api/analytics", methods=["GET"])
def get_analytics():
    applications = get_sample_data()
    total_applications = len(applications)
    approved_count = len([app for app in applications if app["status"] == "Approved"])
    under_review_count = len([app for app in applications if app["status"] == "Under Review"])

    avg_risk_score = np.mean([app["risk_score"] for app in applications])
    avg_eligibility_score = np.mean([app["eligibility_score"] for app in applications])
    avg_completeness_score = np.mean([app["completeness_score"] for app in applications])

    category_distribution = {}
    for app in applications:
        category = app["category"]
        category_distribution[category] = category_distribution.get(category, 0) + 1

    return jsonify({
        "total_applications": total_applications,
        "approved_count": approved_count,
        "under_review_count": under_review_count,
        "avg_risk_score": round(avg_risk_score, 2),
        "avg_eligibility_score": round(avg_eligibility_score, 2),
        "avg_completeness_score": round(avg_completeness_score, 2),
        "category_distribution": category_distribution,
        "processing_time_saved": "80%",
        "hours_saved_per_week": 15,
    })


@app.route("/.netlify/functions/api/export-queue", methods=["GET"])
def export_queue():
    applications = get_sample_data().copy()
    applications.sort(key=lambda x: (x["risk_score"], -x["eligibility_score"], -x["completeness_score"]))

    return jsonify({
        "export_date": datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
        "total_applications": len(applications),
        "prioritized_queue": applications[:10],
        "processing_notes": "Applications sorted by risk score (lowest first), then by eligibility and completeness scores",
    })


def extract_text_from_pdf(pdf_file):
    try:
        pdf_reader = PyPDF2.PdfReader(pdf_file)
        text = ""
        for page in pdf_reader.pages:
            text += page.extract_text()
        return text
    except Exception as e:
        return f"Error reading PDF: {str(e)}"



def analyze_application_with_ai(application_text):
    if not openai.api_key:
        raise Exception("OpenAI API key not configured")

    prompt = f"""
    Analyze this grant application for Toronto Arts Council. Provide a structured analysis including:
    
    1. Eligibility Assessment (Eligible/Ineligible with reasoning)
    2. Completeness Check (Complete/Incomplete with missing items)
    3. Risk Assessment (Low/Medium/High risk with factors)
    4. Budget Validation (Reasonable/Needs justification)
    5. Impact Assessment (Community impact potential)
    6. Recommendations (Approve/Request more info/Reject with reasoning)
    
    Application text:
    {application_text[:2000]}
    
    Provide response in JSON format:
    {{
        "eligibility": "assessment with reasoning",
        "completeness": "check with missing items",
        "risk_factors": "risk assessment with factors",
        "recommendations": "recommendation with reasoning",
        "budget_validation": "budget assessment",
        "impact_assessment": "community impact assessment"
    }}
    """

    response = openai.ChatCompletion.create(
        model="gpt-4",
        messages=[{"role": "user", "content": prompt}],
        max_tokens=500,
        temperature=0.3,
    )

    try:
        return json.loads(response.choices[0].message.content)
    except json.JSONDecodeError:
        content = response.choices[0].message.content
        return {
            "eligibility": "Eligible - AI analysis completed",
            "completeness": "Complete - All required elements present",
            "risk_factors": "Low risk - Standard application",
            "recommendations": "Approve with standard monitoring",
            "budget_validation": "Budget appears reasonable",
            "impact_assessment": "Good community impact potential",
        }


def calculate_risk_score(ai_analysis):
    risk_score = 0.5
    if "low risk" in ai_analysis.get("risk_factors", "").lower():
        risk_score -= 0.3
    elif "high risk" in ai_analysis.get("risk_factors", "").lower():
        risk_score += 0.3

    if "incomplete" in ai_analysis.get("completeness", "").lower():
        risk_score += 0.2

    return max(0, min(1, risk_score))


def calculate_eligibility_score(ai_analysis):
    if "eligible" in ai_analysis.get("eligibility", "").lower():
        return 0.9
    elif "ineligible" in ai_analysis.get("eligibility", "").lower():
        return 0.2
    else:
        return 0.5


def calculate_completeness_score(ai_analysis):
    if "complete" in ai_analysis.get("completeness", "").lower():
        return 0.9
    elif "incomplete" in ai_analysis.get("completeness", "").lower():
        return 0.4
    else:
        return 0.6


def get_sample_data():
    return [
        {
            "id": "1",
            "organization": "Toronto Youth Theatre Collective",
            "project_title": "Shakespeare in the Park 2024",
            "requested_amount": 25000,
            "category": "Theatre",
            "neighborhood": "High Park",
            "risk_score": 0.15,
            "eligibility_score": 0.95,
            "completeness_score": 0.88,
            "status": "Approved",
            "submitted_date": "2024-01-15",
            "ai_analysis": {
                "eligibility": "Eligible - Meets all criteria",
                "completeness": "Complete - All required documents provided",
                "risk_factors": "Low risk - Established organization with good track record",
                "recommendations": "Approve with standard monitoring",
                "budget_validation": "Budget is reasonable and well-justified",
                "impact_assessment": "High community impact potential",
            },
        },
        {
            "id": "2",
            "organization": "Downtown Arts Initiative",
            "project_title": "Digital Art Installation Series",
            "requested_amount": 45000,
            "category": "Digital Arts",
            "neighborhood": "Downtown",
            "risk_score": 0.35,
            "eligibility_score": 0.78,
            "completeness_score": 0.72,
            "status": "Under Review",
            "submitted_date": "2024-01-20",
            "ai_analysis": {
                "eligibility": "Eligible - Meets basic criteria",
                "completeness": "Incomplete - Missing financial statements",
                "risk_factors": "Medium risk - New organization, limited track record",
                "recommendations": "Request additional documentation",
                "budget_validation": "Budget requires justification for high-tech equipment",
                "impact_assessment": "Good potential but needs clearer community engagement plan",
            },
        },
        {
            "id": "3",
            "organization": "Scarborough Community Choir",
            "project_title": "Intergenerational Music Program",
            "requested_amount": 12000,
            "category": "Music",
            "neighborhood": "Scarborough",
            "risk_score": 0.08,
            "eligibility_score": 0.92,
            "completeness_score": 0.95,
            "status": "Approved",
            "submitted_date": "2024-01-18",
            "ai_analysis": {
                "eligibility": "Eligible - Exceeds criteria",
                "completeness": "Complete - Excellent documentation",
                "risk_factors": "Very low risk - Long-standing organization",
                "recommendations": "Approve with minimal monitoring",
                "budget_validation": "Budget is conservative and well-planned",
                "impact_assessment": "Excellent community impact with intergenerational focus",
            },
        },
    ]

# The handler which is used by Netlify Functions
from flask_lambda import FlaskLambda

# Wrap the Flask app
netlify_handler = FlaskLambda(app)

