requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==1.3.7
python-dotenv==1.0.0
PyPDF2==3.0.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
python-multipart==0.0.6
Werkzeug==2.3.7
flask-lambda==0.1.0
